Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main(){
-    float a,b,c,average;
+    int a,b,c,average;
     scanf("%d %d %d",&a,&b,&c);
     average = (a+b+c)/3;
     printf("Average: %.2f",average);
